Add standalone agent support (#104)
This adds a new agent type (`StandaloneAgent`) which can handle queries
directly without any base prompt or fewshots.

The bulk of the change is factoring out most of `CodeShotAgent` into an
`AgentBase` type that is now shared between `CodeShotAgent` and
`StandaloneAgent`.

In addition to supporting registered funcs, `StandaloneAgents` can
accept queries directly at `POST /`.
