+++ fixieai/agents/__init__.py
@@ -1,16 +1,19 @@
+from fixieai.agents.agent_base import AgentBase
 from fixieai.agents.api import AgentQuery
 from fixieai.agents.api import AgentResponse
 from fixieai.agents.api import Embed
 from fixieai.agents.api import Message
 from fixieai.agents.code_shot import CodeShotAgent
-from fixieai.agents.code_shot import LlmSettings
 from fixieai.agents.corpora import DocumentCorpus
 from fixieai.agents.corpora import DocumentLoader
+from fixieai.agents.llm_settings import LlmSettings
 from fixieai.agents.oauth import OAuthHandler
 from fixieai.agents.oauth import OAuthParams
+from fixieai.agents.standalone import StandaloneAgent
 from fixieai.agents.user_storage import UserStorage
 
 __all__ = [
+    "AgentBase",
     "AgentQuery",
     "AgentResponse",
     "CodeShotAgent",
@@ -19,6 +22,7 @@ __all__ = [
     "Embed",
     "LlmSettings",
     "Message",
+    "StandaloneAgent",
     "OAuthHandler",
     "OAuthParams",
     "UserStorage",
